Preserve column ordering (#14)
* Add failing test

* Test still fails, even when `WithColumn0` used

This confirms that we need a new approach to creating the fields in the
dataframe

* Order columns using orderedmap when parsing JSON record

The JSON records that DuckDB dumps are actually ordered maps. The object
keys (and values) appear in the order that they were requested in the
SELECT statement.
Although JSON has no concept of ordering within objects, this is the
simplest and quickest way to fix this problem.

I'll try to investigate some other avenues for fixing this one day in the
future. @ryantxu suggested a couple of things we could try:
- We can implement a JSON parser that goes straight to data frame (like the one in the sdk)
- or better may be using parquet output. if named based on a hash this could also be a cache system 'for free'

Also note, there's another orderedmap Go package we could have used:
https://github.com/wk8/go-ordered-map
They seem to have a similar number of forks and stars on Github, and I
just happened across the iancoleman one first.